Create folder in advance and make subfolder with consignment.
The first issue is that the s3 getObject method which writes to a file
won't write to a directory that doesn't exist so we have to create the
directory first.

The second issue is that we wouldn't want files from different
consignments with the same path and filename (unlikely but possible)
overwriting each other.