Refactor error handling
Pass exceptions around in the form of failed futures or failed tries, rather
than just returning a string message. This means that we will be able to log
the full stack trace when we add a logging framework. It also makes the return
values easier to interpret, because a `Try[String]` has a clear failure case,
whereas it's less obvious what `Either[String, String]` represents.