Convert droid results to FFID metadata class
This means running the droid commands, parsing the csv file and
converting those to the FFIDMetadata class that the API expects.

There is also a fix for when the graphql api returns Some("") for the
original path. We need to check for this and for some reason it stopped
working so this is fixed here with toOptEmpty.

The RecordProcessor code is then changed to deal with the new return
value from the output method.