Add tests for the output method.
There is some logic in this method now where there wasn't for Siegfried
so it could use some tests.

To do this, I've refactored the dummy shell script to take a csv file
name as a parameter which allows the command to output different csvs so
we can test various things.