Refactor to use Droid and s3 downloader.
This looks like quite a major change but this is mostly removing the
code which got the original path from the API and saved the file to EFS.
This has now been moved to another project
https://github.com/nationalarchives/tdr-download-files

This will be run in a lambda attached to the s3 bucket and this will
take care of saving the file so all this lambda has to do is run Droid
against the correct file.

Droid is extremely slow at the moment so we're changing the settings for
the lambda to only allow one message through at once. This may be
changed as Droid changes.