Replace the jdk base image with alpine.
Originally, I copied the jre directory to the EFS volume and used that
to run java.
I realised after that the lambda is running in a java container with
java installed so I don't need to do this.
So all this container has to do is download droid and unzip it.
I've replaced it with alpine so the image size has gone from ~600Mb to
about 40Mb