Update tests to check the original path is correct
Before each test, the file with the original path is written in much the
same way as it would be by the download files lambda. The test script
then checks to see if it is there. If we're passing the original path
quoted, then the path will exist and the test will pass.