Escape double quotes.
We have a similar problem in the download-files lambda where it's been
changed to use the File api but as this is running a droid command
instead of mkidr, this won't work here.

I've kept the double quotes around the file path but I've escaped out
any double quotes in the file. This should allow any special characters
in the files.