Make sure that every before-and-after step is run in tests
Before, when multiple traits were mixed into a test, only one set of
`beforeEach` (or similar) methods was run, because it overrode the
the method from the other mixin.

Add a call to `super` to every before or after method, so that all of
the setup and destroy steps are run.

Follow the ScalaTest advice in https://www.scalatest.org/user_guide/sharing_fixtures
to call `super.beforeEach` at the end of the method and `super.afterEach`
at the start of the method.