Reset message visibility if an error happens when running FFID check
This makes Lambda retries faster. They had become much slower when we
increased the SQS message visibility to be three times the length of the
Lambda timeout: https://github.com/nationalarchives/tdr-dev-documentation/blob/master/architecture-decision-records/0020-sqs-visibility-timeout.md

This change means that the retry happens immediately, not when the message
has expired after 45 minutes. 45 minutes. 45 minutes. 45 minutes.

The message visibility reset only happens if an error is encountered when
running the FFID check. If the error happens before the SQS message is parsed,
we do not have a receipt handle available to reset the visibility.