Remove ErrorSummary wrapper class
ErrorSummary contained a Throwable and a message, which is essentially
just the same as a wrapping exception. So return another Throwable
instead with the same message and cause that were being passed into
ErrorSummary.

This will make it easier to add extra error handling, such as when we
reset the message visibility using the receipt handle, because there
aren't so many layers of errors.