Fail the lambda if there are no droid matches.
Every so often, Droid will return zero matches for a file which should
never happen. It's going to be incredibly difficult to work out why
Droid is doing this. If you do run the files again for the ones which
return no matches then it works so we'll just throw an exception and let
the retries have another go.