Old droid with new architecture
This has been refactored to take in json without the SNS wrapper and to
write the results to the lambda output. This will then be used in the
new step function backend checks.

I also wanted to get rid of EFS so I've found a way of calling the same
APIs which Droid does for its own command line. This is still very
inefficent because it's still creating the profile and we do need the
new Droid changes but this will reduce our dependency on other teams
releases.