Get latest version of signature files
When the lambda starts from cold, it carries out the following steps.

* Checks to see if the /tmp directory contains container-* or
  DROID_Signature*
* If they exist, use those to create the Droid instance.
* If they don't exist:
  * Call http://www.nationalarchives.gov.uk/pronom/container-signature.xml and get the last modified date
  * Convert this to YYYYMMDD and use that for the container signature version
  * Call the SOAP endpoint at http://www.nationalarchives.gov.uk/pronom/service.asmx
  * This returns the latest Droid signature file version.
  * Download both these files from the TNA CDN as before.

I've done it this way because we should only be calling the
nationalarchives url when the lambda cold starts, if we're running a lot
of files at once, it won't get called repeatedly. When the lambda does
cold start though, we'll get the latest version.

If this doesn't work or we end up hitting the TNA website too often,
we'll have to remove this and look at setting up a cron job to update it
periodically but we have enough jobs like that already so I wanted to
try it this way.