Pin the droid version.
There is an issue with the new Droid version 110 which is causing the
file format lambda to fail. We always automatically download the latest
version. This will pin it to 109 which is the last known working version.

I've set an environment variable so that we can stop pinning the version
wiht an environment variable so we can test it and revert it if we need
to.