Make the tests pass
There was a few bits that needed changing.

DROID calls S3 with a range which is passed as a header in the http request.
None of the tests files are that large and the default chunk size is 4K
so I've added two wiremock stubs to cover the calls with the two
headers. These return the correct range of bytes from the test file.

The HEAD request wasn't stubbed so DROID wasn't getting the file size
from the headObject call which means it can't identify properly. I've
added in a stub mapping for that.

The pdf is fmt/18 not fmt/276

The signature files didn't have enough information in them for DROID to
work. For the binary signature file I've added.

* The signature for a zip file so DROID can tell that the docx and xlsx
  are zips and look at the container signature
* The signature for fmt/18 (pdf)
* File format mappings for the docx and xlsx container signatures

The container signature file has:
* Signatures for docx and xlsx
* Associated file format mappings
* A trigger puid for zip so it can find them.

I've added test AWS credentials to the test command because DROID is
using the SDK and it checks for them, even if it doesn't need them.